Cast syscall(SYS_gettid) to uint32_t explicitly to make Coverity happy (#133)

+++ utils/utils.h
@@ -199,7 +199,7 @@ inline uint32_t GetTid() {
   return GetCurrentThreadId();
 #else
 #ifdef SYS_gettid
-  return syscall(SYS_gettid);
+  return (uint32_t)syscall(SYS_gettid);
 #else
   #error "SYS_gettid is unavailable on this system"
 #endif
